Move filter summary to new position
