+++ src/common/components/multiSelectDropdown/dropdownMenu.module.scss
@@ -12,6 +12,7 @@
     top: 100%;
     left: 0;
     right: 0;
+    z-index: 2;
   }
 
   .dropdownMenuWrapper {
