+++ src/domain/eventSearch/filterSummary/searchWordFilter.module.scss
